Adding add books button styles
